ECOPROJECT-4034 | refactor: extract chat history VM and add OpenShift nginx config
