+++ deploy/dev/nginx-openshift.conf
@@ -1,0 +1,92 @@
+# Static nginx config for OpenShift deployments with co-located
+# migration-planner and oma-lightspeed services. Use this when you need
+# hardcoded service names instead of the envsubst-based nginx.conf.template.
+events {
+    worker_connections 1024;
+}
+http {
+    include /etc/nginx/mime.types;
+    default_type application/octet-stream;
+    sendfile on;
+    keepalive_timeout 65;
+    client_max_body_size 50M;
+
+    log_format main '$remote_addr - $remote_user [$time_local] "$request" '
+                      '$status $body_bytes_sent "$http_referer" '
+                      '"$http_user_agent" "$http_x_forwarded_for"';
+
+    upstream migration_planner_api {
+        server migration-planner:3443;
+    }
+
+    upstream oma_lightspeed {
+        server oma-lightspeed:8080;
+    }
+
+    server {
+        listen 8081;
+        server_name localhost;
+        root /usr/share/nginx/html;
+        index index.html;
+
+        # Proxy chat API calls to oma-lightspeed service (streaming support)
+        location /api/chat/ {
+            proxy_pass http://oma_lightspeed/;
+            proxy_http_version 1.1;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+            proxy_set_header Connection '';
+            proxy_connect_timeout 300s;
+            proxy_send_timeout 300s;
+            proxy_read_timeout 300s;
+            proxy_buffering off;
+            proxy_cache off;
+            chunked_transfer_encoding on;
+        }
+
+        # Proxy legacy API path to migration-planner-api service (without rewrite)
+        location /api/migration-assessment/ {
+            proxy_pass http://migration_planner_api;
+            proxy_http_version 1.1;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+            proxy_connect_timeout 300s;
+            proxy_send_timeout 300s;
+            proxy_read_timeout 300s;
+        }
+
+        # Proxy API calls to migration-planner-api service
+        location /api/ {
+            proxy_pass http://migration_planner_api;
+            proxy_http_version 1.1;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+            proxy_connect_timeout 300s;
+            proxy_send_timeout 300s;
+            proxy_read_timeout 300s;
+        }
+
+        # Serve static files and SPA
+        location / {
+            try_files $uri $uri/ /index.html;
+        }
+
+        # Enable gzip compression
+        gzip on;
+        gzip_types text/plain text/css application/json application/javascript text/xml application/xml application/xml+rss text/javascript;
+
+        # Security headers
+        add_header X-Frame-Options "SAMEORIGIN" always;
+        add_header X-Content-Type-Options "nosniff" always;
+        add_header X-XSS-Protection "1; mode=block" always;
+
+        access_log /dev/stdout main;
+        error_log /dev/stderr info;
+    }
+}
